Move PortabilityJob and AuthFlowConfiguration types up into their respective SPI packages
